# tests: default to more threads for ui-tests