@llogiq 2022-02-20
I worry that this may increase the global load on the system to the point where it outweighs any time savings. Our jobs are not alone on the CI machines and someone has to pay for the cycles. Given this, I don't feel confident r+ing this.

@rustbot label +S-needs-discussion