@matthiaskrgr 2022-02-20
Hm, I would expect the host container to have some mechanism to limit the number of threads that get assigned to some arbitrary limit. Otherwise everyone would just `make -j 999` in order to get faster builds :sweat_smile: 

I mostly had users testing locally in mind, I only noticed it while running tests locally while glancing at htop and there I still had some headroom in my cpu utilization which is fixed by this change.