@flip1995 2022-02-21
GitHub actions has a [limited amount of cores](https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources). This may be higher for the Rust org, but I would be surprised if there wasn't a limit at all. You could check how many threads we have available by temporary adding a step to the CI calling `nproc`.

So I'm not concerned that this will destroy the CI. Though even for the local user, using all available threads by default might be a bad idea. So I would limit the default to just half of the available threads, if we should go for parallel-by-default.