@flip1995 2022-02-21
That's true. I see a speedup of `x1.48` on my i7-5820x @ 6 cores / 12 threads. Only using half the threads reduces the speedup to `x1.26`. Since running all those tests doesn't take that long (~10s with multithreads for me), I think it should be ok to use all available threads by default.