@matthiaskrgr 2022-02-21
oh wow, didn't think it would have that much of an impact with more cores! (also reminds me to get a faster machine 😭)