@flip1995 2022-02-21
I just tried running it on our server with 40 cores (12 were 100% in use, so effectively 28 cores). I only got a speedup of `= x1.03` there. 

Ah I just realized: compile-test already uses #cores, but this PR forces it to use #threads instead?