@llogiq 2022-02-21
I worry mostly about the effect on other tenants on the system, but that may be misguided.

In any event, 10% wall time will add up, so I'm not exactly opposed to the PR.