@matthiaskrgr 2022-02-21
>Ah I just realized: compile-test already uses #cores, but this PR forces it to use #threads instead?

Ah right, `num_cpus::get()` will return the number of threads of a system (4 in my case)
https://docs.rs/num_cpus/latest/num_cpus/fn.get.html